requirements.txt
googletrans==4.0.0rc1
python-docx==1.1.2
Werkzeug==3.0.3
flask==3.0.3